RISC-V: testsuite: Skip test with -flto
Hi,

the zbb-rol-ror and stack_save_restore tests use the -fno-lto option and
scan the final assembly.  For an invocation like -flto ... -fno-lto the
output file we scan is still something like
  zbb-rol-ror-09.ltrans0.ltrans.s.

Therefore skip the tests when "-flto" is present.  This gets rid
of a few UNRESOLVED tests.

Regtested on rv64gcv_zvl512b.  Going to push if the CI agrees.

Regards
 Robin

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/stack_save_restore_1.c: Skip for -flto.
	* gcc.target/riscv/stack_save_restore_2.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-04.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-05.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-06.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-07.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-08.c: Ditto.
	* gcc.target/riscv/zbb-rol-ror-09.c: Ditto.