ada: Fix Generate_Minimal_Reproducer on instantiations
Before this patch, the code that creates a copy of the semantic closure
with the default naming convention was incorrect when the compiler was
processing a library unit that was an instantiation of a generic with a
body. This patch adds code to detect that situation and adjusts the
copying process accordingly.

gcc/ada/ChangeLog:

	* generate_minimal_reproducer.adb (Generate_Minimal_Reproducer):
	Fix when main library item is an instantiation.