middle-end: Fix incorrect type replacement in operands_equals [PR118472]
In g:3c32575e5b6370270d38a80a7fa8eaa144e083d0 I made a mistake and incorrectly
replaced the type of the arguments of an expression with the type of the
expression.  This is of course wrong.

This reverts that change and I have also double checked the other replacements
and they are fine.

gcc/ChangeLog:

	PR middle-end/118472
	* fold-const.cc (operand_compare::operand_equal_p): Fix incorrect
	replacement.

gcc/testsuite/ChangeLog:

	PR middle-end/118472
	* gcc.dg/pr118472.c: New test.