[testsuite] rearrange requirements for dfp bitint run tests
dfp.exp sets the default to compile when dfprt is not available, but
some dfp bitint tests override the default without that requirement,
and try to run even when dfprt is not available.

Instead of overriding the default, rewrite the requirements so that
they apply even when compiling, since the absence of bitint or of
int128 would presumably cause compile failures.


for  gcc/testsuite/ChangeLog

	* gcc.dg/dfp/bitint-1.c: Rewrite requirements to retain dfprt.
	* gcc.dg/dfp/bitint-2.c: Likewise.
	* gcc.dg/dfp/bitint-3.c: Likewise.
	* gcc.dg/dfp/bitint-4.c: Likewise.
	* gcc.dg/dfp/bitint-5.c: Likewise.
	* gcc.dg/dfp/bitint-6.c: Likewise.
	* gcc.dg/dfp/bitint-7.c: Likewise.
	* gcc.dg/dfp/bitint-8.c: Likewise.
	* gcc.dg/dfp/int128-1.c: Likewise.
	* gcc.dg/dfp/int128-2.c: Likewise.
	* gcc.dg/dfp/int128-3.c: Likewise.
	* gcc.dg/dfp/int128-4.c: Likewise.