d: Fix record layout of compiler-generated TypeInfo_Class [PR115249]
In r14-8766, the layout of TypeInfo_Class changed in the runtime
library, but didn't get reflected in the compiler-generated data,
causing a corruption of runtime type introspection on BigEndian targets.

This adjusts the size of the `ClassFlags' field from uint to ushort, and
adds a new ushort `depth' field in the space where ClassFlags used to
occupy.

	PR d/115249

gcc/d/ChangeLog:

	* typeinfo.cc (create_tinfo_types): Update internal Typenfo
	representation.
	(TypeInfoVisitor::visit (TypeInfoClassDeclaration *)): Likewise.