c++/modules: Propagate FNDECL_USED_AUTO when propagating deduced return types [PR118049]
In the linked testcase, we're erroring because the declared return types
of the functions do not appear to match.  This is because when merging
the deduced return types for 'foo' in 'auto-5_b.C', we overwrote the
return type for the declaration with the deduced return type from
'auto-5_a.C' but neglected to track that we were originally declared
with 'auto'.

As a drive-by improvement to QOI, also add checks for if the deduced
return types do not match; this is currently useful because we do not
check the equivalence of the bodies of functions yet.

	PR c++/118049

gcc/cp/ChangeLog:

	* module.cc (trees_in::is_matching_decl): Propagate
	FNDECL_USED_AUTO as well.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/auto-5_a.C: New test.
	* g++.dg/modules/auto-5_b.C: New test.
	* g++.dg/modules/auto-5_c.C: New test.
	* g++.dg/modules/auto-6_a.H: New test.
	* g++.dg/modules/auto-6_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>
(cherry picked from commit f054c36c4fcb693e04411dc691ef4172479143d6)